infra/cdk/lib/lambdas/extract_avatars/requirements.txt
playwright==1.42.0
openai==1.99.3
pydantic==2.6.1
boto3
requests
beautifulsoup4
Pillow
